+++ src/smart_follower_control/include/smart_follower_control/follower_runtime.hpp
@@ -16,9 +16,10 @@ struct FollowerRuntimeConfig
 {
   double control_rate{20.0};
   double target_distance{1.0};
-  double steering_kalman_process_noise{6.0};
-  double steering_kalman_measurement_noise{0.02};
-  double steering_kalman_initial_covariance{1.0};
+  double theta_deadzone{0.03};
+  double stop_hold_distance{0.05};
+  double stop_hold_angle{0.10};
+  double stop_hold_speed_mps{0.12};
   double target_timeout{0.4};
   double prediction_horizon_s{0.25};
   double velocity_ema_alpha{0.70};
@@ -50,9 +51,9 @@ struct FollowerRuntimeSnapshot
   double target_speed_mps{0.0};
   double prediction_age_s{-1.0};
   bool predicted_target_valid{false};
-  bool steering_filter_ready{false};
+  bool stale_target_hold{false};
+  bool hold_zone_active{false};
   double raw_theta{0.0};
-  double filtered_theta{0.0};
   int last_pose_lock_id{-1};
   int last_pose_lock_state{0};
   bool locked_track_found{false};
@@ -85,32 +86,20 @@ private:
     bool valid{false};
   };
 
-  struct SteeringKalmanState
-  {
-    bool initialized{false};
-    double theta{0.0};
-    double theta_rate{0.0};
-    double p00{1.0};
-    double p01{0.0};
-    double p10{0.0};
-    double p11{1.0};
-    rclcpp::Time stamp{0, 0, RCL_ROS_TIME};
-  };
-
   void configure_controllers();
   std::optional<TargetState> predict_target(const rclcpp::Time & now_time);
   double rate_limit(double target, double current, double accel_limit, double dt) const;
-  void reset_steering_filter();
-  double filter_theta_measurement(double theta, const rclcpp::Time & stamp);
+  bool should_hold_stop(double distance_error, double raw_theta, const TargetState & target) const;
 
   FollowerRuntimeConfig config_;
   TargetState last_target_;
   PID pid_r_;
   PID pid_t_;
-  SteeringKalmanState steering_filter_;
   geometry_msgs::msg::Twist last_cmd_;
   double last_raw_theta_{0.0};
-  double last_filtered_theta_{0.0};
+  bool stale_target_hold_active_{false};
+  bool last_hold_zone_active_{false};
+  rclcpp::Time last_pose_message_stamp_{0, 0, RCL_ROS_TIME};
   int last_pose_lock_id_{-1};
   int last_pose_lock_state_{0};
   bool last_pose_locked_track_found_{false};